requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
streamlit==1.38.0
openai==1.47.0
python-docx==1.1.2
PyPDF2==3.0.1
pandas==2.2.2
python-dotenv==1.0.1
aiofiles==24.1.0
jinja2==3.1.4
httpx==0.27.2
